% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readal}
\alias{readal}
\title{readal}
\usage{
readal(args = arglist_get(...))
}
\arguments{
\item{args}{Arguments for readal}
}
\description{
Run readal
}
\examples{
library(outsider)
readal <- module_import('readal', repo = 'dombennett/om..trimal')
readal('--help')
}
