% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trimal}
\alias{trimal}
\title{trimal}
\usage{
trimal(args = arglist_get(...))
}
\arguments{
\item{args}{Arguments for trimal}
}
\description{
Run trimal
}
\examples{
library(outsider)
trimal <- module_import('trimal', repo = 'dombennett/om..trimal')
trimal('--help')
}
